pyproject.toml
[tool.poetry]
name = "lidarr-metadata-server"
version = "0.0.0"
description = "Self-hosted replacement for LidarrAPI.Metadata"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "lidarr_metadata_server" },
    { include = "capture_proxy" }
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
httpx = "^0.27.0"
loguru = "^0.7.2"
PyYAML = "^6.0.1"
jsonschema = "^4.23.0"
sqlmodel = "^0.0.16"
aiosqlite = "^0.19.0"
python-dotenv = "^1.0.1"
aiofiles = "^23.2.1"
brotli = "^1.1.0"
psutil = "^6.0.0"
unidecode = "^1.3.8"
jsonpointer = "^2.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5"
pytest-cov = "^5.0.0"
black = "^24.4.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
mypy = "^1.10.0"
pre-commit = "^3.5.0"
types-aiofiles = "^24.1.0.20250708"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
